Finished e1

+++ DS-61-03-A2/src/e1/Staff.java
@@ -4,16 +4,15 @@ import java.lang.reflect.Member;
 
 abstract public class Staff extends Members{
 
-    float baseSalary;
+    int baseSalary;
 
     public Staff(String memberName, String memberSurname, int age, int horrocruxCount) {
         super(memberName, memberSurname, age, horrocruxCount);
     }
 
-    abstract float getSalary();
+    abstract int getSalary();
 
     String printSalary(){
-
         return memberName + " " + memberSurname
                 + " ( " +  catString() + " ): "
                 + getSalary() + " galleons";
